pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "vtoroe_dyhanie"
version = "1.0.0"
description = "Телеграм-бот для автоматизации HR-процессов фонда Второе Дыхание"
authors = [
    "Голодных Евгений",
    "Квас Евгений",
    "Киланов Андрей",
    "Ковалева Мария",
    "Корепанов Дмитрий",
    "Латышева Виктория",
    "Максимов Владимир",
    "Мамонов Александр",
    "Мартынова Настасья",
    "Мольков Андрей",
    "Насибуллин Дмитрий",
    "Савельева Анастасия",
    "Евгений Коваленко",
    "Александр Новожилов",
    "Олег Исаев",
    "Михаил Волков",
    "Дмитрий Белолипецкий",
]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
poetry = "1.6.1"
python-telegram-bot = {extras = ["webhooks"], version = "^20.7"}
pydantic = "2.3.0"
requests = "2.26.0"
pydantic-settings = "2.0.3"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
autopep8 = "2.0.4"
flake8-isort = ["6.1.0"]
mypy = "^1.6.1"
django = "4.2.7"
django-jazzmin = "2.6.0"
aiohttp = "3.8.6"
sqlalchemy-utils = "^0.41.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.extras]
flake8 = ["3.9.2"]
isort = ["5.12.0"]
python-dotenv = ["0.19.0"]
autopep8 = ["2.11.0"]
add-trailing-comma = ["3.1.0"]
mypy = ["1.5.1"]
aiosqlite = ["0.19.0"]

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"

[tool.flake8]
max-line-length = "79"
paths = ["./bot", "./core", "./src"]
max-complexity = "10"
exclude = ["tests", ".venv", "venv"]
per-file-ignores = [
    "__init__.py:F401",
]

[tool.isort]
default_section = "THIRDPARTY"
sections = ["STDLIB", "FIRSTPARTY", "THIRDPARTY", "LOCALFOLDER"]
known_local_folder = ["src", "bot", "core"]
lines_after_imports = "2"
skip = ["tests", ".venv", "venv"]
line_length = "79"
multi_line_output = "3"
include_trailing_comma = "true"
use_parentheses = "true"

[tool.black]
line-length = 79
skip-string-normalization = "true"
